dockerfiles reorganisations
