+++ dockerfiles/dockerfile.models
@@ -1,0 +1,12 @@
+# Utiliser une image officielle Python comme base
+FROM python:3.10-slim
+
+# Définir le répertoire de travail à l'intérieur du conteneur
+WORKDIR /app
+
+# Installer les dépendances
+COPY ./requirements/models.txt /app/requirements.txt
+RUN pip install --no-cache-dir -r /app/requirements.txt
+
+# Commande pour exécuter l'API
+CMD ["uvicorn", "src.api.main:app", "--host", "0.0.0.0", "--port", "8000"]
