+++ requirements/data.txt
@@ -1,0 +1,3 @@
+pandas
+scikit-learn
+requests
