+++ requirements/requirements.txt
@@ -1,4 +1,3 @@
-
 requests==2.31.0
 awscli>=1.29.0
 python-dotenv>=0.5.1
@@ -7,3 +6,4 @@ uvicorn
 pandas
 scikit-learn
 click
+numpy
