--- src/models/evaluate/requirements.txt
@@ -1,2 +1,0 @@
-pandas
-numpy
