--- src/models/train/requirements.txt
@@ -1,2 +1,0 @@
-pandas
-scikit-learn
